Package script and update docs
